switch to prod test project network etc.
