+++ gce_ovf_import_tests/test_suites/ovf_instance_import/ovf_import_test_suite.go
@@ -1,467 +1,474 @@
-////  Copyright 2019 Google Inc. All Rights Reserved.
-////
-////  Licensed under the Apache License, Version 2.0 (the "License");
-////  you may not use this file except in compliance with the License.
-////  You may obtain a copy of the License at
-////
-////      http://www.apache.org/licenses/LICENSE-2.0
-////
-////  Unless required by applicable law or agreed to in writing, software
-////  distributed under the License is distributed on an "AS IS" BASIS,
-////  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-////  See the License for the specific language governing permissions and
-////  limitations under the License.
+//  Copyright 2019 Google Inc. All Rights Reserved.
 //
-//// Package ovftestsuite contains e2e tests for OVF import
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+// Package ovfinstanceimporttestsuite contains e2e tests for instance import cli tools
 package ovfinstanceimporttestsuite
 
-//
-//import (
-//	"context"
-//	"fmt"
-//	"io/ioutil"
-//	"log"
-//	"os"
-//	"regexp"
-//	"strings"
-//	"sync"
-//	"time"
-//
-//	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/paramhelper"
-//	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/path"
-//	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/gce_ovf_import/ovf_import_params"
-//	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/gce_ovf_import/ovf_importer"
-//	daisyCompute "github.com/GoogleCloudPlatform/compute-image-tools/daisy/compute"
-//	"github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/compute"
-//	computeUtils "github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/compute"
-//	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/junitxml"
-//	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/test_config"
-//	"github.com/kylelemons/godebug/pretty"
-//	api "google.golang.org/api/compute/v1"
-//)
-//
-//const (
-//	testSuiteName = "OVFImportTests"
-//)
-//
-//var (
-//	dump = &pretty.Config{IncludeUnexported: true}
-//)
-//
-//var vf = func(
-//	inst *compute.Instance, vfString string, port int64, interval, timeout time.Duration) error {
-//	return inst.WaitForSerialOutput(vfString, port, interval, timeout)
-//}
-//
-//type ovfImportTestProperties struct {
-//	importParams          *ovfimportparams.OVFImportParams
-//	name                  string
-//	isWindows             bool
-//	description           string
-//	startup               *api.MetadataItems
-//	assertTimeout         time.Duration
-//	vf                    func(*compute.Instance, string, int64, time.Duration, time.Duration) error
-//	expectedMachineType   string
-//	expectedStartupOutput string
-//}
-//
-//// TestSuite is OVF import test suite.
-//func TestSuite(
-//	ctx context.Context, tswg *sync.WaitGroup, testSuites chan *junitxml.TestSuite,
-//	logger *log.Logger, testSuiteRegex, testCaseRegex *regexp.Regexp,
-//	testProjectConfig *testconfig.Project) {
-//
-//	defer tswg.Done()
-//
-//	if testSuiteRegex != nil && !testSuiteRegex.MatchString(testSuiteName) {
-//		return
-//	}
-//
-//	testSuite := junitxml.NewTestSuite(testSuiteName)
-//	defer testSuite.Finish(testSuites)
-//	suffix := path.RandString(5)
-//	ovaBucket := "compute-image-tools-test-resources"
-//	logger.Printf("Running TestSuite %q", testSuite.Name)
-//	region, _ := paramhelper.GetRegion(testProjectConfig.TestZone)
-//
-//	startupScriptUbuntu3disks := loadScriptContent(
-//		"gce_ovf_import_tests/scripts/ovf_import_test_ubuntu_3_disks.sh", logger)
-//	startupScriptLinuxSingleDisk := loadScriptContent(
-//		"daisy_integration_tests/scripts/post_translate_test.sh", logger)
-//	startupScriptWindowsSingleDisk := loadScriptContent(
-//		"daisy_integration_tests/scripts/post_translate_test.ps1", logger)
-//	startupScriptWindowsTwoDisks := loadScriptContent(
-//		"gce_ovf_import_tests/scripts/ovf_import_test_windows_two_disks.ps1", logger)
-//
-//	testSetup := []*ovfImportTestProperties{
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-ubuntu-3-disks-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/ubuntu-1604-three-disks", ovaBucket),
-//				OsID:          "ubuntu-1604",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-ubuntu-three-disks-%s", suffix),
-//			description: "Ubuntu 3 disks, one data disk larger than 10GB.",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"startup-script", startupScriptUbuntu3disks),
-//			assertTimeout:         7200 * time.Second,
-//			expectedStartupOutput: "All tests passed!",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-centos-6-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/", ovaBucket),
-//				OsID:          "centos-6",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-centos-6-%s", suffix),
-//			description: "Centos 6.8",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"startup-script", startupScriptLinuxSingleDisk),
-//			assertTimeout:         7200 * time.Second,
-//			expectedMachineType:   "n1-standard-4",
-//			expectedStartupOutput: "All tests passed!",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-w2k12-r2-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/w2k12-r2", ovaBucket),
-//				OsID:          "windows-2012r2",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-8",
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-w2k12-r2-%s", suffix),
-//			isWindows:   true,
-//			description: "Windows 2012 R2 two disks",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"windows-startup-script-ps1", startupScriptWindowsTwoDisks),
-//			assertTimeout:         7200 * time.Second,
-//			expectedMachineType:   "n1-standard-8",
-//			expectedStartupOutput: "All Tests Passed",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-w2k16-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/w2k16/w2k16.ovf", ovaBucket),
-//				OsID:          "windows-2016",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n2-standard-2",
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-w2k16-%s", suffix),
-//			isWindows:   true,
-//			description: "Windows 2016",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"windows-startup-script-ps1", startupScriptWindowsSingleDisk),
-//			assertTimeout:         7200 * time.Second,
-//			expectedMachineType:   "n2-standard-2",
-//			expectedStartupOutput: "All Tests Passed",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-w2k8r2-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/win2008r2-all-updates-four-nic.ova", ovaBucket),
-//				OsID:          "windows-2008r2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-w2k8r2-%s", suffix),
-//			isWindows:   true,
-//			description: "Windows 2008r2 - Four NICs",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"windows-startup-script-ps1", startupScriptWindowsSingleDisk),
-//			assertTimeout:         7200 * time.Second,
-//			expectedStartupOutput: "All Tests Passed",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-debian-9-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/bitnami-tomcat-8.5.43-0-linux-debian-9-x86_64.ova", ovaBucket),
-//				OsID:          "debian-9",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//			},
-//			// no startup script as this OVA has issues running it (possibly due to no SSH allowed)
-//			// b/141321520
-//			name:                  fmt.Sprintf("ovf-import-test-debian-9-%s", suffix),
-//			description:           "Debian 9",
-//			expectedMachineType:   "n1-standard-4",
-//			expectedStartupOutput: "All tests passed!",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-instance-virtualbox-6-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/ubuntu-16.04-virtualbox.ova", ovaBucket),
-//				OsID:          "ubuntu-1604",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-virtualbox-%s", suffix),
-//			description: "Ubuntu 1604 from Virtualbox",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"startup-script", startupScriptLinuxSingleDisk),
-//			assertTimeout:         7200 * time.Second,
-//			expectedMachineType:   "n1-standard-4",
-//			expectedStartupOutput: "All tests passed!",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("aws-ova-ubuntu-1604-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/ova/aws-ova-ubuntu-1604.ova", ovaBucket),
-//				OsID:          "ubuntu-1604",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//			},
-//			name:        fmt.Sprintf("aws-ova-ubuntu-1604-%s", suffix),
-//			description: "Ubuntu 1604 from AWS",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-network-name-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/", ovaBucket),
-//				OsID:          "centos-6",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//				Network:       fmt.Sprintf("%v-vpc-1", testProjectConfig.TestProjectID),
-//				Subnet:        fmt.Sprintf("%v-subnet-1", testProjectConfig.TestProjectID),
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-network-name-%s", suffix),
-//			description: "Test network setting (name only)",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"startup-script", startupScriptLinuxSingleDisk),
-//			assertTimeout:         7200 * time.Second,
-//			expectedMachineType:   "n1-standard-4",
-//			expectedStartupOutput: "All tests passed!",
-//		},
-//		{
-//			importParams: &ovfimportparams.OVFImportParams{
-//				ClientID:      "test",
-//				InstanceNames: fmt.Sprintf("test-network-path-%v", suffix),
-//				OvfOvaGcsPath: fmt.Sprintf("gs://%v/", ovaBucket),
-//				OsID:          "centos-6",
-//				Labels:        "lk1=lv1,lk2=kv2",
-//				Project:       &testProjectConfig.TestProjectID,
-//				Zone:          testProjectConfig.TestZone,
-//				MachineType:   "n1-standard-4",
-//				Network:       fmt.Sprintf("global/networks/%v-vpc-1", testProjectConfig.TestProjectID),
-//				Subnet:        fmt.Sprintf("projects/%v/regions/%v/subnetworks/%v-subnet-1", testProjectConfig.TestProjectID, region, testProjectConfig.TestProjectID),
-//			},
-//			name:        fmt.Sprintf("ovf-import-test-network-path-%s", suffix),
-//			description: "Test network setting (path)",
-//			startup: computeUtils.BuildInstanceMetadataItem(
-//				"startup-script", startupScriptLinuxSingleDisk),
-//			assertTimeout:         7200 * time.Second,
-//			expectedMachineType:   "n1-standard-4",
-//			expectedStartupOutput: "All tests passed!",
-//		},
-//	}
-//
-//	var wg sync.WaitGroup
-//	tests := make(chan *junitxml.TestCase)
-//	for _, setup := range testSetup {
-//		wg.Add(1)
-//		go ovfImportTestCase(ctx, setup, tests, &wg, logger, testCaseRegex, testProjectConfig)
-//	}
-//
-//	go func() {
-//		wg.Wait()
-//		close(tests)
-//	}()
-//
-//	for ret := range tests {
-//		testSuite.TestCase = append(testSuite.TestCase, ret)
-//	}
-//
-//	logger.Printf("Finished TestSuite %q", testSuite.Name)
-//}
-//
-//func ovfImportTestCase(
-//	ctx context.Context, testSetup *ovfImportTestProperties, tests chan *junitxml.TestCase,
-//	wg *sync.WaitGroup, logger *log.Logger, regex *regexp.Regexp,
-//	testProjectConfig *testconfig.Project) {
-//
-//	defer wg.Done()
-//
-//	ovfImportTestCase := junitxml.NewTestCase(
-//		testSuiteName, fmt.Sprintf("[OVFImport] %v", testSetup.description))
-//
-//	for tc, f := range map[*junitxml.TestCase]func(
-//		context.Context, *junitxml.TestCase, *ovfImportTestProperties, *log.Logger, *testconfig.Project){
-//		ovfImportTestCase: runOvfImportTest,
-//	} {
-//		if tc.FilterTestCase(regex) {
-//			tc.Finish(tests)
-//		} else {
-//			logger.Printf("Running TestCase %s.%q", tc.Classname, tc.Name)
-//			f(ctx, tc, testSetup, logger, testProjectConfig)
-//			tc.Finish(tests)
-//			logger.Printf("TestCase %s.%q finished in %fs", tc.Classname, tc.Name, tc.Time)
-//		}
-//	}
-//}
-//
-//func runOvfImportTest(
-//	ctx context.Context, testCase *junitxml.TestCase, testSetup *ovfImportTestProperties,
-//	logger *log.Logger, testProjectConfig *testconfig.Project) {
-//
-//	logger.Printf("[%v] Creating OVF importer", testSetup.name)
-//	ovfImporter, err := ovfimporter.NewOVFImporter(testSetup.importParams)
-//	ovfImporter.BuildID = testSetup.name
-//
-//	if err != nil {
-//		logger.Printf("[%v] error creating OVF importer: %v", testSetup.name, err)
-//		testCase.WriteFailure("error creating OVF importer: %v", err)
-//		return
-//	}
-//
-//	logger.Printf("[%v] Starting OVF import", testSetup.name)
-//	_, err = ovfImporter.Import()
-//	if err != nil {
-//		logger.Printf("[%v] error while performing OVF import: %v", testSetup.name, err)
-//		testCase.WriteFailure("error while performing OVF import: %v", err)
-//		ovfImporter.CleanUp()
-//		return
-//	}
-//
-//	logger.Printf("[%v] OVF import finished", testSetup.name)
-//	ovfImporter.CleanUp()
-//
-//	// assertion
-//	client, err := daisyCompute.NewClient(ctx)
-//	if err != nil {
-//		testCase.WriteFailure("Error creating client: %v", err)
-//		return
-//	}
-//
-//	instanceName := testSetup.importParams.InstanceNames
-//
-//	instance, err := client.GetInstance(
-//		testProjectConfig.TestProjectID, testProjectConfig.TestZone, instanceName)
-//	if err != nil {
-//		testCase.WriteFailure("Error retrieving instance `%v` in `%v` zone: %v", instanceName,
-//			testProjectConfig.TestZone, err)
-//		return
-//	}
-//
-//	instanceWrapper := computeUtils.Instance{Instance: instance, Client: client,
-//		Project: testProjectConfig.TestProjectID, Zone: testProjectConfig.TestZone}
-//
-//	// The boot disk for a Windows instance must have the WINDOWS GuestOSFeature,
-//	// while the boot disk for other operating systems shouldn't have it.
-//	for _, disk := range instance.Disks {
-//		if !disk.Boot {
-//			continue
-//		}
-//
-//		hasWindowsFeature := false
-//		for _, feature := range disk.GuestOsFeatures {
-//			if "WINDOWS" == feature.Type {
-//				hasWindowsFeature = true
-//				break
-//			}
-//		}
-//
-//		if testSetup.isWindows && !hasWindowsFeature {
-//			testCase.WriteFailure(
-//				"Windows boot disk missing WINDOWS GuestOsFeature. Features found=%v",
-//				disk.GuestOsFeatures)
-//		} else if !testSetup.isWindows && hasWindowsFeature {
-//			testCase.WriteFailure(
-//				"Non-Windows boot disk includes WINDOWS GuestOsFeature. Features found=%v",
-//				disk.GuestOsFeatures)
-//		}
-//	}
-//
-//	defer func() {
-//		logger.Printf("[%v] Deleting instance `%v`", testSetup.name, instanceName)
-//		instanceWrapper.Cleanup()
-//	}()
-//
-//	if !strings.HasSuffix(instance.MachineType, testSetup.expectedMachineType) {
-//		testCase.WriteFailure(
-//			"Instance machine type `%v` doesn't match the expected machine type `%v`",
-//			instance.MachineType, testSetup.expectedMachineType)
-//		return
-//	}
-//
-//	if !strings.HasSuffix(instance.Zone, testSetup.importParams.Zone) {
-//		testCase.WriteFailure("Instance zone `%v` doesn't match requested zone `%v`",
-//			instance.Zone, testSetup.importParams.Zone)
-//		return
-//	}
-//
-//	logger.Printf("[%v] Stopping instance before restarting with test startup script", testSetup.name)
-//	err = client.StopInstance(
-//		testProjectConfig.TestProjectID, testProjectConfig.TestZone, instanceName)
-//
-//	if err != nil {
-//		testCase.WriteFailure("Error stopping imported instance: %v", err)
-//		return
-//	}
-//
-//	if testSetup.startup == nil {
-//		logger.Printf("[%v] Will not set test startup script to instance metadata as it's not defined", testSetup.name)
-//		return
-//	}
-//
-//	logger.Printf("[%v] Setting instance metadata with test startup script", testSetup.name)
-//	err = client.SetInstanceMetadata(testProjectConfig.TestProjectID, testProjectConfig.TestZone,
-//		instanceName, &api.Metadata{Items: []*api.MetadataItems{testSetup.startup},
-//			Fingerprint: instance.Metadata.Fingerprint})
-//
-//	if err != nil {
-//		testCase.WriteFailure("Couldn't set instance metadata to verify OVF import: %v", err)
-//		return
-//	}
-//
-//	logger.Printf("[%v] Starting instance with test startup script", testSetup.name)
-//	err = client.StartInstance(
-//		testProjectConfig.TestProjectID, testProjectConfig.TestZone, instanceName)
-//	if err != nil {
-//		testCase.WriteFailure("Couldn't start instance to verify OVF import: %v", err)
-//		return
-//	}
-//
-//	logger.Printf("[%v] Waiting for `%v` in instance serial console.", testSetup.name,
-//		testSetup.expectedStartupOutput)
-//	if err := instanceWrapper.WaitForSerialOutput(
-//		testSetup.expectedStartupOutput, 1, 5*time.Second, 15*time.Minute); err != nil {
-//		testCase.WriteFailure("Error during VM validation: %v", err)
-//		return
-//	}
-//}
-//
-//func loadScriptContent(scriptPath string, logger *log.Logger) string {
-//	scriptContent, err := ioutil.ReadFile(scriptPath)
-//	if err != nil {
-//		logger.Printf("Error loading script `%v`: %v", scriptPath, err)
-//		os.Exit(1)
-//	}
-//	return string(scriptContent)
-//}
+import (
+	"context"
+	"fmt"
+	"io/ioutil"
+	"log"
+	"os"
+	"regexp"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/paramhelper"
+	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/path"
+	daisyCompute "github.com/GoogleCloudPlatform/compute-image-tools/daisy/compute"
+	computeUtils "github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/compute"
+	ovftestsuite "github.com/GoogleCloudPlatform/compute-image-tools/gce_ovf_import_tests/test_suites"
+	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/junitxml"
+	"github.com/GoogleCloudPlatform/compute-image-tools/go/e2e_test_utils/test_config"
+)
+
+const (
+	testSuiteName = "OVFImportTests"
+	ovaBucket     = "compute-image-tools-test-resources"
+)
+
+var (
+	cmds = map[ovftestsuite.TestType]string{
+		ovftestsuite.Wrapper:                   "./gce_ovf_import",
+		ovftestsuite.GcloudProdWrapperLatest:   "gcloud",
+		ovftestsuite.GcloudLatestWrapperLatest: "gcloud",
+	}
+)
+
+type ovfImportTestProperties struct {
+	instanceName              string
+	isWindows                 bool
+	expectedStartupOutput     string
+	verificationStartupScript string
+	zone                      string
+	sourceUri                 string
+	os                        string
+	machineType               string
+	network                   string
+	subnet                    string
+}
+
+// TestSuite is image import test suite.
+func TestSuite(
+	ctx context.Context, tswg *sync.WaitGroup, testSuites chan *junitxml.TestSuite,
+	logger *log.Logger, testSuiteRegex, testCaseRegex *regexp.Regexp,
+	testProjectConfig *testconfig.Project) {
+
+	testsMap := map[ovftestsuite.TestType]map[*junitxml.TestCase]func(
+			context.Context, *junitxml.TestCase, *log.Logger, *testconfig.Project, ovftestsuite.TestType){}
+
+	testTypes := []ovftestsuite.TestType{
+		ovftestsuite.Wrapper,
+		ovftestsuite.GcloudProdWrapperLatest,
+		ovftestsuite.GcloudLatestWrapperLatest,
+	}
+	for _, testType := range testTypes {
+		instanceImportUbuntu3DisksTestCase := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Ubuntu 3 disks, one data disk larger than 10GB"))
+		instanceImportCentos68 := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Centos 6.8"))
+		instanceImportWindows2012R2TwoDisks := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Windows 2012 R2 two disks"))
+		instanceImportWindows2016 := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Windows 2016"))
+		instanceImportWindows2008R2FourNICs := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Windows 2008r2 - Four NICs"))
+		instanceImportDebian9 := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Debian 9"))
+		instanceImportUbuntu16FromVirtualBox := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Ubuntu 1604 from Virtualbox"))
+		instanceImportUbuntu16FromAWS := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Ubuntu 1604 from AWS"))
+		instanceImportNetworkSettingsName := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Test network setting (name only)"))
+		instanceImportNetworkSettingsPath := junitxml.NewTestCase(
+			testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", testType, "Test network setting (path)"))
+
+		testsMap[testType] = map[*junitxml.TestCase]func(
+			context.Context, *junitxml.TestCase, *log.Logger, *testconfig.Project, ovftestsuite.TestType){}
+		testsMap[testType][instanceImportUbuntu3DisksTestCase] = runOVFInstanceImportUbuntu3Disks
+		testsMap[testType][instanceImportCentos68] = runOVFInstanceImportCentos68
+		testsMap[testType][instanceImportWindows2012R2TwoDisks] = runOVFInstanceImportWindows2012R2TwoDisks
+		testsMap[testType][instanceImportWindows2016] = runOVFInstanceImportWindows2016
+		testsMap[testType][instanceImportWindows2008R2FourNICs] = runOVFInstanceImportWindows2008R2FourNICs
+		testsMap[testType][instanceImportDebian9] = runOVFInstanceImportDebian9
+		testsMap[testType][instanceImportUbuntu16FromVirtualBox] = runOVFInstanceImportUbuntu16FromVirtualBox
+		testsMap[testType][instanceImportUbuntu16FromAWS] = runOVFInstanceImportUbuntu16FromAWS
+		testsMap[testType][instanceImportNetworkSettingsName] = runOVFInstanceImportNetworkSettingsName
+		testsMap[testType][instanceImportNetworkSettingsPath] = runOVFInstanceImportNetworkSettingsPath
+	}
+
+	//testsMap[ovftestsuite.Wrapper] = map[*junitxml.TestCase]func(
+	//		context.Context, *junitxml.TestCase, *log.Logger, *testconfig.Project, ovftestsuite.TestType){}
+	//
+	//testsMap[ovftestsuite.Wrapper][junitxml.NewTestCase(
+	//	testSuiteName, fmt.Sprintf("[%v][OVFInstanceImport] %v", ovftestsuite.Wrapper, "Windows 2016"))] = runOVFInstanceImportWindows2016
+
+	ovftestsuite.TestSuite(ctx, tswg, testSuites, logger, testSuiteRegex, testCaseRegex,
+		testProjectConfig, testSuiteName, testsMap)
+}
+
+func runOVFInstanceImportUbuntu3Disks(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-ubuntu-3-disks-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"gce_ovf_import_tests/scripts/ovf_import_test_ubuntu_3_disks.sh", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All tests passed!",
+		sourceUri:             fmt.Sprintf("gs://%v/ova/ubuntu-1604-three-disks", ovaBucket),
+		os:                    "ubuntu-1604",
+		machineType:           "n1-standard-4"}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportCentos68(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-centos-6-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All tests passed!",
+		sourceUri:             fmt.Sprintf("gs://%v/", ovaBucket),
+		os:                    "centos-6",
+		machineType:           "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportWindows2012R2TwoDisks(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-w2k12-r2-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"gce_ovf_import_tests/scripts/ovf_import_test_windows_two_disks.ps1", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All Tests Passed",
+		sourceUri:             fmt.Sprintf("gs://%v/ova/w2k12-r2", ovaBucket),
+		os:                    "windows-2012r2",
+		machineType:           "n1-standard-8",
+		isWindows:             true,
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportWindows2016(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-w2k16-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.ps1", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All Tests Passed",
+		sourceUri:             fmt.Sprintf("gs://%v/ova/w2k16/w2k16.ovf", ovaBucket),
+		os:                    "windows-2016",
+		machineType:           "n2-standard-2",
+		isWindows:             true,
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportWindows2008R2FourNICs(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-w2k8r2-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.ps1", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All Tests Passed",
+		sourceUri:             fmt.Sprintf("gs://%v/ova/win2008r2-all-updates-four-nic.ova", ovaBucket),
+		os:                    "windows-2008r2",
+		isWindows:             true,
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportDebian9(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	// no startup script as this OVA has issues running it (possibly due to no SSH allowed)
+	// b/141321520
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-debian-9-%v", suffix),
+		zone:         testProjectConfig.TestZone,
+		sourceUri:    fmt.Sprintf("gs://%v/ova/bitnami-tomcat-8.5.43-0-linux-debian-9-x86_64.ova", ovaBucket),
+		os:           "debian-9",
+		machineType:  "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportUbuntu16FromVirtualBox(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-virtualbox-6-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All tests passed!",
+		sourceUri:             fmt.Sprintf("gs://%v/ova/ubuntu-16.04-virtualbox.ova", ovaBucket),
+		os:                    "ubuntu-1604",
+		machineType:           "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportUbuntu16FromAWS(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-instance-aws-ova-ubuntu-1604-%v", suffix),
+		zone: testProjectConfig.TestZone,
+		sourceUri:   fmt.Sprintf("gs://%v/ova/aws-ova-ubuntu-1604.ova", ovaBucket),
+		os:          "ubuntu-1604",
+		machineType: "n1-standard-4",
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportNetworkSettingsName(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-network-name-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All tests passed!",
+		sourceUri:             fmt.Sprintf("gs://%v/", ovaBucket),
+		os:                    "centos-6",
+		machineType:           "n1-standard-4",
+		network:               fmt.Sprintf("%v-vpc-1", testProjectConfig.TestProjectID),
+		subnet:                fmt.Sprintf("%v-subnet-1", testProjectConfig.TestProjectID),
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func runOVFInstanceImportNetworkSettingsPath(ctx context.Context, testCase *junitxml.TestCase, logger *log.Logger,
+	testProjectConfig *testconfig.Project, testType ovftestsuite.TestType) {
+
+	suffix := path.RandString(5)
+	region, _ := paramhelper.GetRegion(testProjectConfig.TestZone)
+	props := &ovfImportTestProperties{
+		instanceName: fmt.Sprintf("test-network-path-%v", suffix),
+		verificationStartupScript: loadScriptContent(
+			"daisy_integration_tests/scripts/post_translate_test.sh", logger),
+		zone:                  testProjectConfig.TestZone,
+		expectedStartupOutput: "All tests passed!",
+		sourceUri:             fmt.Sprintf("gs://%v/", ovaBucket),
+		os:                    "centos-6",
+		machineType:           "n1-standard-4",
+		network:               fmt.Sprintf("global/networks/%v-vpc-1", testProjectConfig.TestProjectID),
+		subnet:                fmt.Sprintf("projects/%v/regions/%v/subnetworks/%v-subnet-1", testProjectConfig.TestProjectID, region, testProjectConfig.TestProjectID),
+	}
+
+	runOVFInstanceImportTest(ctx, buildTestArgs(props, testProjectConfig)[testType], testType, testProjectConfig, logger, testCase, props)
+}
+
+func buildTestArgs(props *ovfImportTestProperties, testProjectConfig *testconfig.Project) map[ovftestsuite.TestType][]string {
+	gcloudArgs := []string{
+		"beta", "compute", "instances", "import", props.instanceName, "--quiet",
+		"--docker-image-tag=latest",
+		fmt.Sprintf("--project=%v", testProjectConfig.TestProjectID),
+		fmt.Sprintf("--source-uri=%v", props.sourceUri),
+		fmt.Sprintf("--zone=%v", testProjectConfig.TestZone),
+	}
+	wrapperArgs := []string{"-client-id=e2e", fmt.Sprintf("-project=%v", testProjectConfig.TestProjectID),
+		fmt.Sprintf("-instance-names=%s", props.instanceName),
+		fmt.Sprintf("-ovf-gcs-path=%v", props.sourceUri),
+		fmt.Sprintf("-zone=%v", testProjectConfig.TestZone),
+	}
+
+	if props.os != "" {
+		gcloudArgs = append(gcloudArgs, fmt.Sprintf("--os=%v", props.os))
+		wrapperArgs = append(wrapperArgs, fmt.Sprintf("-os=%v", props.os))
+	}
+	if props.machineType != "" {
+		gcloudArgs = append(gcloudArgs, fmt.Sprintf("--machine-type=%v", props.machineType))
+		wrapperArgs = append(wrapperArgs, fmt.Sprintf("-machine-type=%v", props.machineType))
+	}
+	if props.network != "" {
+		gcloudArgs = append(gcloudArgs, fmt.Sprintf("--network=%v", props.network))
+		wrapperArgs = append(wrapperArgs, fmt.Sprintf("-network=%v", props.network))
+	}
+	if props.subnet != "" {
+		gcloudArgs = append(gcloudArgs, fmt.Sprintf("--subnet=%v", props.subnet))
+		wrapperArgs = append(wrapperArgs, fmt.Sprintf("-subnet=%v", props.subnet))
+	}
+
+	argsMap := map[ovftestsuite.TestType][]string{
+		ovftestsuite.Wrapper:                   wrapperArgs,
+		ovftestsuite.GcloudProdWrapperLatest:   gcloudArgs,
+		ovftestsuite.GcloudLatestWrapperLatest: gcloudArgs,
+	}
+	return argsMap
+}
+
+func runOVFInstanceImportTest(ctx context.Context, args []string, testType ovftestsuite.TestType,
+	testProjectConfig *testconfig.Project, logger *log.Logger, testCase *junitxml.TestCase,
+	props *ovfImportTestProperties) {
+
+	if ovftestsuite.RunTestForTestType(cmds[testType], args, testType, logger, testCase) {
+		verifyImportedInstance(ctx, testCase, testProjectConfig, logger, props)
+	}
+}
+
+func failure(testCase *junitxml.TestCase, logger *log.Logger, msg string) {
+	testCase.WriteFailure(msg)
+	logger.Printf(msg)
+}
+
+func verifyImportedInstance(
+	ctx context.Context, testCase *junitxml.TestCase, testProjectConfig *testconfig.Project,
+	logger *log.Logger, props *ovfImportTestProperties) {
+
+	client, err := daisyCompute.NewClient(ctx)
+	if err != nil {
+		failure(testCase, logger, fmt.Sprintf("Error creating client: %v", err))
+		return
+	}
+
+	logger.Printf("Verifying imported instance...")
+	instance, err := computeUtils.CreateInstanceObject(ctx, testProjectConfig.TestProjectID, props.zone, props.instanceName, props.isWindows)
+	if err != nil {
+		failure(testCase, logger, fmt.Sprintf("Image '%v' doesn't exist after import: %v", props.instanceName, err))
+		return
+	}
+
+	defer func() {
+		logger.Printf("Deleting instance `%v`", props.instanceName)
+		if err := instance.Cleanup(); err != nil {
+			logger.Printf("Instance '%v' failed to clean up: %v", props.instanceName, err)
+		} else {
+			logger.Printf("Instance '%v' cleaned up.", props.instanceName)
+		}
+	}()
+
+	// The boot disk for a Windows instance must have the WINDOWS GuestOSFeature,
+	// while the boot disk for other operating systems shouldn't have it.
+	for _, disk := range instance.Disks {
+		if !disk.Boot {
+			continue
+		}
+
+		hasWindowsFeature := false
+		for _, feature := range disk.GuestOsFeatures {
+			if "WINDOWS" == feature.Type {
+				hasWindowsFeature = true
+				break
+			}
+		}
+
+		if props.isWindows && !hasWindowsFeature {
+			testCase.WriteFailure(
+				"Windows boot disk missing WINDOWS GuestOsFeature. Features found=%v",
+				disk.GuestOsFeatures)
+		} else if !props.isWindows && hasWindowsFeature {
+			testCase.WriteFailure(
+				"Non-Windows boot disk includes WINDOWS GuestOsFeature. Features found=%v",
+				disk.GuestOsFeatures)
+		}
+	}
+
+	if props.machineType != "" && !strings.HasSuffix(instance.MachineType, props.machineType) {
+		testCase.WriteFailure(
+			"Instance machine type `%v` doesn't match the expected machine type `%v`",
+			instance.MachineType, props.machineType)
+		return
+	}
+
+	if !strings.HasSuffix(instance.Zone, props.zone) {
+		failure(testCase, logger, fmt.Sprintf("Instance zone `%v` doesn't match requested zone `%v`",
+			instance.Zone, props.zone))
+		return
+	}
+
+	logger.Printf("[%v] Stopping instance before restarting with test startup script", props.instanceName)
+	err = client.StopInstance(
+		testProjectConfig.TestProjectID, props.zone, props.instanceName)
+
+	if err != nil {
+		testCase.WriteFailure("Error stopping imported instance: %v", err)
+		return
+	}
+
+	if props.verificationStartupScript == "" {
+		logger.Printf("[%v] Will not set test startup script to instance metadata as it's not defined", props.instanceName)
+		return
+	}
+
+	err = instance.StartWithScript(props.verificationStartupScript)
+	if err != nil {
+		testCase.WriteFailure("Error starting instance `%v` with script: %v", props.instanceName, err)
+		return
+	}
+	logger.Printf("[%v] Waiting for `%v` in instance serial console.", props.instanceName,
+		props.expectedStartupOutput)
+	if err := instance.WaitForSerialOutput(
+		props.expectedStartupOutput, 1, 5*time.Second, 15*time.Minute); err != nil {
+		testCase.WriteFailure("Error during VM validation: %v", err)
+	}
+}
+
+func loadScriptContent(scriptPath string, logger *log.Logger) string {
+	scriptContent, err := ioutil.ReadFile(scriptPath)
+	if err != nil {
+		logger.Printf("Error loading script `%v`: %v", scriptPath, err)
+		os.Exit(1)
+	}
+	return string(scriptContent)
+}
+
+func contains(arr []string, subarr []string) bool {
+	for item := range subarr {
+		exists := false
+		for i := range arr {
+			if item == i {
+				exists = true
+				break
+			}
+		}
+		if !exists {
+			return false
+		}
+	}
+	return true
+}
